input

+++ tok.py
@@ -3,10 +3,6 @@ from pathlib import Path
 from third_party.Botok.botok import Text, sentence_tokenizer, WordTokenizer
 
 
-infile = Path("input/articletest.txt")
-raw = infile.read_text(encoding="utf-8")
-
-
 def sent_tok(raw):
     w = WordTokenizer()
     tokens = w.tokenize(raw, spaces_as_punct=True)
@@ -21,9 +17,16 @@ def plaintext_sent_par(units, sep="\n") -> str:
     return sep.join(out)
 
 
-t = Text(raw).custom_pipeline("basic_cleanup", sent_tok, "dummy", plaintext_sent_par)
+inpath = Path("input/")
+infiles = list(inpath.rglob("*.txt"))
+for f in infiles:
+    raw = f.read_text(encoding="utf-8")
 
-
-outfile = Path("output") / infile.name
-
-outfile.write_text(t, encoding="utf-8")
+    outpath = Path(  "output/sentences") / f.parts[-2]
+    outpath.mkdir(exist_ok=True)
+    outfile = outpath / f.name
+    if not outfile.is_file():
+        print(outfile)
+        t = Text(raw).custom_pipeline("basic_cleanup", sent_tok, "dummy", plaintext_sent_par)
+        print()
+        outfile.write_text(t, encoding="utf-8")
